Solution using the Dictionary class added
The Dictionary class was added to improve runtime complexity. The time complexity of the Dicitonary solution is O(n) since we only go through the for loop once and each time we look at the table it only costs O(1). The space complexity is O(n) due to the extra storage requirement being dependent on how many items are stored in the hash table.